+++ RELEASE_NOTES.md
@@ -7,6 +7,10 @@ Requested by @octomiku01
 - Closing Instance works if it is your own instance
 - Shows Age Gated instances across all modals and lists
 Requested by @octomiku01
+* "Remember Window Size" - in Settings > Window Behavior
+Requested by @aghostofthepast
+* "Remember Monitor" - in Settings > Window Behavior
+Requested by @kivvio
 
 ## Instance List
 * Redesigned the instance list design to match the new V2 design.